cambios y mejoras en la organizacion del servidor y el cliente
